some cleaning

+++ FGEngine/MyGame/Projectile.cpp
@@ -152,7 +152,7 @@ void Projectile::DrawColliderCircle()
 	for (int i = 0; i < samples; i++)
 	{
 		SDL_RenderDrawLine(camera->GetInternalRenderer(),
-			positions[i].x, positions[i].y, positions[i + 1].x, positions[i + 1].y);
+			(int)positions[i].x, (int)positions[i].y, (int)positions[i + 1].x, (int)positions[i + 1].y);
 	}
 
 	SDL_SetRenderDrawColor(camera->GetInternalRenderer(), 0, 0, 0, 255);
